fix: resolve schema conflict by moving type definitions and restoring complete schema
Root cause analysis and fix:
- DetectionMethodModel and ConfigMappingModel were defined in application_resource.go
- GetDetectionMethodsSchemaBlock() and GetConfigMappingsSchemaBlock() in application_detection_models.go couldn't access them
- This caused compilation errors, so schema blocks were removed earlier
- Tests expected these blocks to exist in the application resource schema

Solution:
- Moved DetectionMethodModel and ConfigMappingModel to application_detection_models.go
- Added missing types import to application_detection_models.go
- Restored detection_methods and config_mappings blocks to application resource schema
- Verified generate job works without conflicts
- TestApplicationResource/Schema now passes locally

This resolves the original schema conflict from first principles.
